requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
mysqlclient==2.2.1
pydantic==2.6.0
pydantic_core==2.16.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.51
typing_extensions==4.9.0
Werkzeug==3.0.1
